fix: Ensure float32 dtype for chemical descriptors
Fixed RuntimeError during training by explicitly converting numpy arrays to float32 when creating torch tensors. Added .float() calls to fingerprint and descriptor tensors to ensure consistent dtype throughout the model pipeline.

Issue: Chemical descriptors were being converted from numpy float64 to torch Double, while model layers expected Float (float32).
Solution: Explicit .float() conversion in encode_compound() method.

+++ backend/services/genomepath/chemical_encoder.py
@@ -131,13 +131,13 @@ class ChemicalEncoder:
             }
         
         # Get descriptors
-        descriptors = self.get_molecular_descriptors(smiles) if self.use_features else np.zeros(7)
+        descriptors = self.get_molecular_descriptors(smiles) if self.use_features else np.zeros(7, dtype=np.float32)
         if descriptors is None:
             descriptors = np.zeros(7, dtype=np.float32)
         
         return {
-            'fingerprint': torch.from_numpy(fp),
-            'descriptors': torch.from_numpy(descriptors),
+            'fingerprint': torch.from_numpy(fp).float(),
+            'descriptors': torch.from_numpy(descriptors).float(),
             'is_valid': torch.tensor(1.0)
         }
